pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-math-agent"
version = "0.1.0"
description = "Sistema multi-agente per generare riassunti HTML da PDF matematici"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["pdf", "math", "agent", "ai", "summarization", "latex"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "groq>=0.4.1",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "pymupdf>=1.23.0",
    "pdfplumber>=0.10.0",
    "pdf2image>=1.16.3",
    "pytesseract>=0.3.10",
    "pix2tex>=0.1.2",
    "datapizza>=0.1.0",
    "jinja2>=3.1.3",
    "markdown>=3.5.0",
    "rich>=13.7.0",
    "click>=8.1.7",
    "tqdm>=4.66.0",
    "html5lib>=1.1",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "flake8>=6.1.0",
    "mypy>=1.8.0",
    "isort>=5.13.0",
]

[project.scripts]
pdf-math-agent = "src.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/pdf-math-agent"
Documentation = "https://github.com/yourusername/pdf-math-agent/blob/main/README.md"
Repository = "https://github.com/yourusername/pdf-math-agent"
Issues = "https://github.com/yourusername/pdf-math-agent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.tox
  | build
  | dist
  | __pycache__
  | outputs
  | checkpoints
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = [
    "tests/",
    "outputs/",
    "checkpoints/",
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
